requirements.txt
asgiref==3.2.10
attrs==21.4.0
boto3==1.20.41
botocore==1.23.41
certifi==2021.10.8
charset-normalizer==2.0.10
coverage==6.3.2
Django==3.1
django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.6
django-filter==21.1
django-js-asset==2.0.0
django-modeltranslation==0.17.5
django-mptt==0.13.4
django-resized==0.3.11
django-rosetta==0.9.8
django-session-timeout==0.1.0
django-sms==0.5.0
django-storages==1.12.3
django-unique-slugify==1.1
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
jmespath==0.10.0
packaging==21.3
Pillow==9.0.0
pluggy==1.0.0
polib==1.1.1
psycopg2-binary==2.9.3
py==1.11.0
pyparsing==3.0.7
pytest==7.0.1
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
requests==2.27.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.8
